beta!!!

+++ plugin/src/main/java/com/crypteam/solana/models/account/AccountInfoValue.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class AccountInfoValue {
     public List<String> data;
     public boolean executable;
-    public int lamports;
+    public long lamports;
     public String owner;
     public int rentEpoch;
 }
